Restore persistence for memcached
This was available for the memcache extension, but when we moved to memcached it went away since the API is different and required changes. The issue with persitence in memcached is that persistent connections require a common identifier, which shouldn't be fixed. Therefore, we need to change a bit the way the memcache servers are configured in config.php as well.