Check explicitly whether array element is defined
Check whether http://macedir.org/entity-category is set in the EntityAttributes
before using it as argument in in_array(). This prevents a "TypeError:
in_array() expects parameter 2 to be array, null given" error.

This situation happens when we do have EntityAttributes in the metadata but
none of type "http://macedir.org/entity-category". An example is an IdP
supporting https://refeds.org/category/research-and-scholarship, which would
have "http://macedir.org/entity-category-support" but not per se also
"http://macedir.org/entity-category".