saml:SP: do not swallow more specific exceptions.
When looking up the IdP to authenticate, the try/catch block would
swallow the more specific exceptions that are generated and display
them all as a single 'not found' exception. This means that the
more specific "Metadata not found" exception page was never rendered,
or also the "Metadata expired" exception not displayed. We don't
need the try/catch here, instead can rely on the exceptions that this
call will generate by itself.