build-release: Use same version number format as in Git
Our releases published to Packagist use version numbers with a
"v"-prefix, e.g. "v1.18.7".

The build-release script overwrites this with a version number without
the "v"-prefix, e.g. "1.18.7".

The result of this is that the composer.json in our release tar-files
does not match the composer.lock file in the repository nor the
composer.json that Packagist looks at.

Fix this by prepending a "v" to the version number when updating
composer.json.