Fix cron invocation from cron by specifying string parameter
Calling the cron module from cron in the adverstised way does
not add an output parameter explicitly. It defaulted to null
in the routing, however the controller does not accept null,
only string for third parameter so one would get a fatal error.
Fix it by setting a string as default.