Document that you must set acs.Bindings if you set AssertionConsumerService to a custom array.
The current metadata generation code will first calculate the supported
protocols, then override the ACS'es with the custom array if given. This
makes that the list of supported protocols may no longer be correct.

The nicest fix is if the supported protocols would be calculated also
taking custom ACS specification in account. This however requires
significant refactor of the sp metadata production which is too invasive
for this bugfix release, especially because the problem is solved
differently in master (and there's no test coverage in 1.19 for this
part). Documenting what you need to do is the safest solution here.