We only support SAML 2.0, so clean up tracking which protocol we support
This auth source no longer supports SAML 1.1 so the ACS code was keeping
track of protocol support with only SAML 2.0 as an option to be added.
Therefore, simplify and set this globally in the class. Keep the outside
interface for now the same where getSupportedProtocols returns as list
of the supported protocol.

Note: if you specify AssertionConsumerService in your config you are
responsible for it having actually bindings that we support. This has
always been the case. This PR makes no attempt to change it (the
specified URLs can even be outside of SSP).