Fix filterscopes using hostname when explicit scopes set.
FilterScopes documented behaviour is that hostname is only used as
a fallback when there are no explicit scopes (makes sense). However
the code does not do that but always includes the hostname to match.

This refactors the code so that hostname is only evaluated when
no explicit scopes are there. Also it takes the whole fetching and
parsing of the SSOlocation outside of the inner loop whcih is
inefficient. This is now only done once, and only if needed (when
no explicit scopes present).

To confirm the changes testcases have been added for this and also
for some other cases to reach 100% coverage for the filter.