getEntityDisplayName to get entityID's name for display
Reimplement getPreferredTranslation outside of IdPDisco so it's reusable
and much more complete. The new method will search the current, default
and fallback language for a displayable string in various metadata
fields. This should be much more complete and robust than previous
approches. Make it a Twig filter so it can also be used there.