Change logging date formatting string from strftime() to date().
strftime is deprecated and will be removed in PHP 9. SSP 2.0 is our
chance to change this format as it's exposed to installers of SSP.

The approch is to keep it simple and just change the format to PHP's
built in date() function which remains supported. Also we do not try
to localize the logging date/timestamps anymore. This matches the syslog
approach that this format tried to mimic. We are using the day without
leading zeroes now also to be a bit closer to how syslog does it.