Fix regexp to remove date from start of string in syslog handler.
Date needs to be removed since this is implicit in syslog. For safety,
remove it only when at the start of the string (default format).
If someone creates a custom string they can leave out date entirely
if it's not at the start of the string.

Being more specific fixes that before elements of the actual log
message were eaten by this regexp if they contained a '%'.

Closes: #1549