Build a 'full' tarball next to the basic tarball.
In 2.0 many modules have been moved to external repositories and can
be installed when needed. There's been some requests to also keep
offering a 'fat' tarball which includes many modules ready to use
by default. This change proposes to build such a tarball as part of
the build process.

The list of modules to include is up for debate, currently more or
less based on what was included in 1.19.