Remove base64attributes feature
This would encode all attributes sent by the IdP (or decode received
by the SP) with base64. As far as I could deduce this is from the
early days of federation, used nowhere anymore, no one came forward
that did, but cannot get certainty. Best way to find out might just
be to drop it. We can easily reinstate it if some use case does turn
up.

Note that the IdP can still perform this behaviour with the more
generic attributeencodings feature.