Error reporting: indicate the actual level instead of always 'Error'
Previously, (just) a warning would be logged as
"Error 2 - something"
"Error 8192 - use of thing is deprecated"

Now this is logged as
"Warning - something"
"Deprecated - use of thing is deprecated"

This should avoid red herrings when debugging an issue when e.g. a
warning, notice or even deprecation notice looks like its an
actual error while it's not.