Fix loginpage_links feature.
It seems with the addition of twig templates another way to process
these links (or other links) has been started but which is broken. The
original way still remains in the code. Make it clearer which links this
concerns by making the template variable more explicit and remove the
unused tabs code.

Closes: #1770