Correct host in generated URLS for IdPs with 'host' config in admin/federation
It gets quite fiddly to get this right. The actual metadata (hosted on
the `host` itself) is alreafy correct because the URL generator uses the
'current' entity metadata and current URL. The admin interface of course
presents all entities in one page and hence cannot rely on the current
URL. Therefore we need to override the url host for this specific
display case.

Closes: #1774