Do not drop UIInfo from SSP array style metadata
It seems likely to be copied from the 1.x way of generating SSP metadata
which also has this removal of UIInfo from the array from the very first
commit: 132a0dd. But it seems safe to just remove the unset since we do add
UIInfo for IdP generated array form metadata.

Closes: #1808