Integrate sanitycheck hook into admin module
The sanitycheck module overlaps the config check functionality in the
admin module. Make it one place where the results of all sanity checks
are shown. This is done in admin so the user does not need to install
a separate module for this - that will no doubt lead to most checks
never being ran.

The change kept the existing sanitycheck hook format, so all modules
implementing this hook will work out of the box without changes
required.

The 'core' module hooks have been removed since the checks duplicate
already present checks in the config page.

As an example, the saml-specific checks have been moved from the admin
to saml module.

More changes are possible, e.g. the check for LDAP could be moved
to the ldap module so it's only shown if you actually use LDAP. Etc.