Workaround for Date header with duplicated value being issued
RunnableResponses would generate a response to the client with
this date header:

  Date: Wed, 23 Aug 2023 11:39:32 GMT, Wed, 23 Aug 2023 11:39:32 GMT

Reproducible with e.g.:
  curl -i https://<ssp base>/saml2/idp/SSOService.php|grep ^Date

I have not been able to exactly pinpoint what causes this duplication.
However, removing it before we pass the response off for further
processing will solve the problem. Since RunnableResponse is already
gone in master, it seems acceptable to add this workaround for 2.0/2.1
because it will by definition have a limited lifetime.

Thanks to David Groep (NIKHEF) for suggesting the workaround.