Back to the two failures that were there before
The presence of a "Default binding" on the below page would seem
to indicate that if the binding is not present in the Array of arrays
then it will be set to the default:

https://github.com/simplesamlphp/simplesamlphp/blob/simplesamlphp-2.2/docs/simplesamlphp-metadata-endpoints.md

The test suite was expecting to be able to do this, though it was also
using the old direct single string representation.

It probably makes sense if we are moving to on the the array of arrays
syntax to either not show a default on the
simplesamlphp-metadata-endpoints page or to use the default if the
binding is not explicitly supplied.