Matches the old outcomes now.
Note that there are some heuristics that seem to have been implicit.

If there is no binding set then one is set for you. This matches the
default value table from
https://github.com/simplesamlphp/simplesamlphp/blob/simplesamlphp-2.2/docs/simplesamlphp-metadata-endpoints.md

If there is only one end point in the array and the binding is not
valid then it is refreshed for you to something that is valid.

On the other hand, for a list of end points no such refreshing is done
so that the case of a non default Location with a valid Binding can
be the one that is used.