A new config option to merge into headers.security for https only
I added getHeadersSecurity() to deal with the merging as this is read
in a few places. A new method seems like a good idea for a security
header rather than trying to special case inside the getOptionalArray() for
headers.security.

This will allow HSTS to be only used when https is in use. It may
break configurations that mix http and https but that is somewhat the
point of HSTS anyway. For testing one might set
headers.security.https.merge to an empty array to allow mixing
http/https in the redirection chain. For production one might like to
force HSTS to avoid accidential http requests.