Allow custom ErrorCodes subclasses through Login::handleLogin nt
This is a version of
https://github.com/simplesamlphp/simplesamlphp/pull/2181
without the test suite update.

I found a few side issues while trying to get the test suite update
going so thought I would split out the code code here for
consideration.

The goal is to allow a custom ErrorCodes subclasses to be used and
have Login keep that association and be able to present information
from those custom error codes on the web page.

This is an update to
https://github.com/simplesamlphp/simplesamlphp/pull/2181

This relates to and should resolve issue
https://github.com/simplesamlphp/simplesamlphp/issues/2133

The ErrorCodes subclass name needs to be saved in the state
information in order to set data['errorcodes'] to the custom error
code class later in the login method.

To restrict which classes can be used this way
Login::registerErrorCodeClass() explicitly tracks each errorcodes
subclass. This is automatic in the constructor of ErrorCodes but it
does mean that any subclass will need to be instantiated at least once
before use or an explicit call to Login::registerErrorCodeClass made.

Although it is not strictly Module related I have put the
resolveNonModuleClass() method in the module class for now to
compliment the resolveClass() method. One needs the have the NonModule
version to allow inline classes and the like to resolve as they
contain the class name followed by a colon and number/offset
information. That colon will make the resolveClass() assume we are
dealing with a module:class specification.