Fixes to encryption optional
The new feature from f99217b43 only worked if a signing cert was
present. With no certs present, getPublicKeys threw an exception.
Fixed the call from encryptAssertion to getPublicKeys to not
require that a key be returned.

Also fixed logic in getPublicKeys
in Configuration.php. Previous logic would return an empty array if
a use type was passed in but no keys of that type were found even if
required was set to true. Thus, the encryptAssertion bug didn't appear
until we had an SP with no certs at all rather than just one without an
encryption cert.