Update the SP and IdP URIs to pass the URI regex.
These were all failing because of
  "eugeneIdP" is not a SAML2-compliant URI

It seems that adding in the colon makes these a compliant test.

The regex that was failing things is
  /^([a-z][a-z0-9\+\-\.]+[:])/i

which is not anchored at both ends, so it seems you can have whatever
you want after the colon. As long as there is one with a limited
prefix.