New 'identifyingAttribute option for SAML Attribute NameID filter (#2260)
* New 'identifyingAttribute option for SAML Attribute NameID filter

Allow a new option in place of identifyingAttribute called
identifyingAttributes. This option expects an array of attribute
names. The first one in the list that's available in the attributes being
released to the SP will be used. This is particularly useful in the hosted
IdP configuration. One can specify multiple attribute names in order of
precedence, and the IdP will release the first one from that list to the
SP that's in the SP's 'attributes' array. This avoids having to define
a new authproc filter for each SP that needs a different value than the
single one that was previously specified using 'identifyingAttribute'.

The change still supports identifyingAttribute option which will be used
instead of the new option if both are set.

* Throw error when AttributeNameID filter has both identifyingAttribute and identifyingAttributes set

* identifyingAttributes clean-up per PR comments

* Cleaned up logic for finding a value for the nameID
* Corrected bug when $value was empty at end of loop

* add a test for the AttributeNameID using an array of attributes

* lint

* Update AttributeNameIDTest.php

These entity ids were copied from the test that I then updated which lead to much better values for test idp and sp entities. I am updating these entities here inline with the comment at https://github.com/simplesamlphp/simplesamlphp/pull/2349#pullrequestreview-2494659186

* Update nameid.md

Mention that these are mutex options.

* Update nameid.md

lint

---------

Co-authored-by: Ben Martin <monkeyiq@users.sourceforge.net>
Co-authored-by: monkeyiq <monkeyiq@gmail.com>