Also allow hosted SP to be defined in metadata/saml20-sp-hosted.php
This was raised in issue
https://github.com/simplesamlphp/simplesamlphp/issues/2299

This is a first attempt to make this happen. If you are defining your
hosted SP in authsources.php then they should still be picked up.
Otherwise you can now define them in metadata/saml20-sp-hosted.php.

This separates the authsources and sp-hosted into two logical
locations. The authsources is a good place to define your
authentication mechanisms that an IdP might like to use to verify a
user. The sp-hosted metadata file is a good place to define a hosted
SP. I think the simplesamlphp-sp.md becomes a little verbose with both
examples all the time. Perhaps we should mention that there are two
formats and the inline definition in authsources is still respected
but that folks might like to move to using sp-hosted. New users might
like to just start using the later as the location that a hosted SP
will be defined.

One slight difference in the format to defining an SP in authsources
vs in sp-hosted is that the type is forced to 'saml:SP' authmatically
for you in sp-hosted. No initial empty 'saml:SP' key is needed there.

I still have to experiment with the setting of the entityid and
entityID for an SP loaded from authsources.php. I suspect that one of
those settings needs to be either not there or made optional.

I have expanded SourceTest.php to try to find issues with the code
though have only tested that with my local setup and not the one that
is employed by phpunit. There are likely to be slight differences to
count()s needed with the later environment.

I initially had getAuthSourcesFromConfig toward the top of the file
but the diff also conflated that method with getSourcesOfType which
was much harder to read. The current location makes the diff much
simpler to read.

As the above highlights, there are probably updates desired to this PR
before merge but I think it is a good time to share it and gather
thoughts and comments.