Use print_r to output the filters to handle cycles (#2368)
* Use print_r to output the filters to handle cycles

Using print_r will allow __debugInfo to customize how things are
print/dumped. This should allow for cases with cycles to be handled
without running out of memory.

Relates to issues/2367

I also found that although this is a Logger::debug the string being
passed to the function is probably being evaluated all the time. Even
if debug is not on the export, dump, or print_r would run. To help
with this I have the debugActive() method to only run the static call
if debug is active. I have some testing using a getString() function
and that will always be called when used as part of the string concat
to the Logger::debug() method call. As the dumping could be expensive
and is not used at all if not debugging this might be handy to have.

Formatting needs a phpcs or whatnot and the new methods in Logger need
some docs. This is mainly an initial drop to discuss.

* phpcs lint

* Remove the debugActive to another PR