A new module to allow debugging SP logins
The ability to test SP logins by non admin users was available in SSP
1.x. In the 2.x series only the admin can do such testing in the admin
module.

This module allows testing SP logins by non admin users. At the moment
there is no login needed in order to test an SP.

This is based on the code from
https://github.com/simplesamlphp/simplesamlphp/pull/2372

I have not brought forward the translations that the code was using in
the admin module. I will add that as a subsequent update once we know
we want this module and where we are wanting to place it.

I have left the list of SP available. Probing for an SP that does not
exist is not an error, it just shows the list again. Trying to use a
non SP also just shows a list of valid SPs that you might try. This
way you can not try to guess the name of an auth mechanism as attempts
at both non existing and non SP return the same result.