A new module to allow debugging SP logins (#2381)
* A new module to allow debugging SP logins

The ability to test SP logins by non admin users was available in SSP
1.x. In the 2.x series only the admin can do such testing in the admin
module.

This module allows testing SP logins by non admin users. At the moment
there is no login needed in order to test an SP.

This is based on the code from
https://github.com/simplesamlphp/simplesamlphp/pull/2372

I have not brought forward the translations that the code was using in
the admin module. I will add that as a subsequent update once we know
we want this module and where we are wanting to place it.

I have left the list of SP available. Probing for an SP that does not
exist is not an error, it just shows the list again. Trying to use a
non SP also just shows a list of valid SPs that you might try. This
way you can not try to guess the name of an auth mechanism as attempts
at both non existing and non SP return the same result.

* Remove copy of status page.

This came from the status.twig in the admin module. Tim mentioned that
there is a base page auth_status.twig which we could use here instead
of making another duplicate. Thanks!

Perhaps we could also look at comparing @admin/status.twig and the
base auth_status.twig page.

* lint

* lint

* Change page title. Add en/fr locales with strings

To add a language you need to put in place a fairly empty file
such as debugsp/locales/fr/LC_MESSAGES/debugsp.po

then
php bin/translations translations:update:translatable

will extract the translation strings into that file for you to translate.

* The CSS from here doesn't seem to do anything for this module.

Might as well not maintain two of them then.

* Might as well just add these po files to make it easier for folks

No need to run any commands to make them in order to translate. The
easier it can be made to add translations for a person who speaks two
langauges the better it might be for SSP :)

* Add a blank comment line for the linter

* Add some docs

* lint

* only a few little psalm issues. including here.