Clear caches script and hook (#2410)
I am not 100% sure about the saml module enable. I thought I would err
on the side of caution leaving it in from the cherry pick.

* Clear caches script and hooke

* Use symofony cache:clear for each module and core

* Overrirde CacheClear Symfony Command

* Github Actions: Create the cache directory.

* Fixing simplesamlphp cache directory

* Fixing chmod permissions

* Fix phpcs issues

* Fix indentation

* config.php.dist file

* config.php.dist file

* Fix failing test after adding the config.php file

* Fix failing test after adding the config.php file

* Clear more caches during update.

Clear the symfony cache to make sure that the ssp-clear command can be found

Using the name "clear-cache" for this script also gave the warning
```
 A script named clear-cache would override a Composer command and has been skipped
```

* Update the docs to be more of a strong recommendation

If we are playing with the cache from composer we really want to
encourage the user to setup that var directory the way that symfony
are wanting it to be setup.

* Include .php.dist files in the sniffer

* Revert previous

---------

Co-authored-by: Tim van Dijen <tvdijen@gmail.com>
Co-authored-by: Ben Martin <monkeyiq@users.sourceforge.net>