Disable clearing the cache for now on update (#2416)
Some installs and composer checkouts are likely to not have the
setfacl in place for the cache. The documentation is updated to
mention that this is a really good idea but in the shorter term we
should not assume this and not run these commands that depend on it
automatically.