Scrub the expire key for session_set_cookie_params (#2426)
* Scrub the expire key for session_set_cookie_params

If that is set then the session_set_cookie_params() will complain. It
is probably best to convert it to working on lifetime but only if
lifetime is also not explicitly set.

* I think if expires is set it should override the lifetime

It is probably more expected if expires is set to 0 or 1 (soon) then
it will be expected to stomp on the lifetime to expire soon.

* Fix typos in the comment.