Update for cookie deletion on logout.
Because a bridge setup runs the IdP and SP (that you are bridging too)
on the same installation instance then cookie handling is a little
more complicated. The SP should not delete the cookie on logout here
as it is shared with the IdP that is bridging to the SP. For that
there is a new saml:SPBridge type that can be used which will be the
same as the normal SP for now but does not delete the session cookie
on logout. For a bridge the IdP and SP are on the same instance so the
IdP can be left to delete the session cookie for that SSP install
instance.

See simplesamlphp-advancedfeatures.md for information about the saml:SPBridge.

This relates to https://github.com/simplesamlphp/simplesamlphp/issues/2435

This PR also has the IdP deleting its cookie on logout. This can be
turned off with the logout_delete_session_cookie (`true` by default)
setting for the IdP. This setting allows degenerate installations with
the IdP and SP on the same instance to continue to function. It is
also a good backup mechanism allowing the new IdP cookie deletion to
be turned off if it is causing issues on a bridge or other setup
instance where further investigation may be needed.

Somehow there were different git hashes there and instead of merging I
rolled back my 2.4 branch and moved it forward. So that should be the
same as my v2 branch but with the current remote history.