Allow the fallback to the pk/cert as mentioned in the docs
This was raised in
https://github.com/simplesamlphp/simplesamlphp/issues/2447

When you have a hosted IdP with
```
    'metadata.sign.enable' => true,
```

This PR will allow the privatekey, certificate to be used for signing
the metadata.

One case that is not handled at the moment is if
metadata.sign.enable is set in the config.php and you are wanting the
privatekey and certificate to be used from the hosted IdP file. If
metadata.sign.enable is set in config.php then unless
you set metadata.sign.enable for that hosted file then things will
fail to work (the IdP doesn't show up in the admin/federation page).

This is sort of inline with the restructions made explicit in
https://github.com/simplesamlphp/simplesamlphp/pull/2466/files