Docs: remove information about fallback from md.sign.pk to pk
Remove information in simplesamlphp-advancedfeatures.md about
possible fallback from metadata.sign.privatekey to privatekey in IdP
and SP metadata files

This was raised in
https://github.com/simplesamlphp/simplesamlphp/issues/2447 as not
working for IdP metadata. I have seen that is currently the case.

It should also be confirmed that the fallback is not actively working
for SP metadata to make sure that this PR does not change existing
functionality.

Perhaps also we should strenghten this to make sure that if
metadata.sign.enable=true in the IdP/SP metadata file then
we *require* you to set the metadata.sign.privatekey and
metadata.sign.certificate. Though that would be a breaking change as
well so should be a 2.5 type udpate.