This goes back to calculating the tag input as before for non module
If we are in a module then the composer.lock is considered to see if a
module might have been updated and thus might have an updated asset.

The tag in master is taken from the mtime just as it was before the
PR.

If the module is being updated with composer it might be a good
optimization to have a composer hook update some cache with the md5 of
the composer lock file so we can directly use that instead of
calculating it from the file.