Enhance_assert_tag (#2490)
* Make tag hash more secure.Modules and SSP core should generate different tags.

* use a specific tagsalt

* change introduced secretsalt to tagsalt

* change that to tagsalt too

* Config rename, use asset file for metadata rather than config.php

Rename tagsalt to assets.secretsalt
For releases shift the target from the config.php to the asset file (either md5 or mtime)

* Move tests to using updated assets.secretsalt config

* Simplify tag input to the mtime/content of the asset itself

* Move to assets.secretsalt for the new asset salt

I may have missed an easy way to talk the config array to find
something. I added getOptionalStringNested to allow assets.secretsalt
to be fetched in the same way that it could be done if it was just a
foo.bar.baz single key in the config file.

* lint

* Add a base64_url_encode() to do rfc 4648 style base64+mods

I am not attached to the name or location. It seems like a good idea
to put this into a function though in case we want to reuse it and
maybe tweak it a bit.

* Remove getOptionalStringNested and directly access the key instead

* Rename assets.secretsalt to assets.salt

* do not touch Configuration.php any more

* This goes back to calculating the tag input as before for non module

If we are in a module then the composer.lock is considered to see if a
module might have been updated and thus might have an updated asset.

The tag in master is taken from the mtime just as it was before the
PR.

If the module is being updated with composer it might be a good
optimization to have a composer hook update some cache with the md5 of
the composer lock file so we can directly use that instead of
calculating it from the file.

* Use hash_file for each tag

The tag encoding is back to 16 values per byte. To combat that I have
increased the tag length to 32 (half the hmac).

* Leave the tag unchanged

That can be updated in a subsequent PR.

A tag length of 5 on a 16 value octet gives a 1 in a million chance of
arbitrary collision.

* more docs for the new salt

* change that use

* remove old tag calculation

* remove blank line

* lint

* docs

---------

Co-authored-by: Ben Martin <monkeyiq@users.sourceforge.net>