Consider it an error if session.check_function can not be called
If the user configuration includes a definition of the check_function
we should not silently ignore it if we can not call it. Consider for
example if the admin is using that to reject or drop a session for a
group of nefarious users. They would rather know that the function is
not callable than assume everything is ok. This could happen after
initial testing for example if the function is in a module that was
not enabled during an upgrade etc.