Consider it an error if session.check_function can not be called (#2498)
* Consider it an error if session.check_function can not be called

If the user configuration includes a definition of the check_function
we should not silently ignore it if we can not call it. Consider for
example if the admin is using that to reject or drop a session for a
group of nefarious users. They would rather know that the function is
not callable than assume everything is ok. This could happen after
initial testing for example if the function is in a module that was
not enabled during an upgrade etc.

* changelog

* Raise an error if the check_function is given but is not callable

Since this is optional, if there is an issue calling the funciton we
should stop right there. The admin has set a security related function
to call and we should consider it an error if something has changed
and the function can not be called.

* lint