Allow a whitelist of errors to be shown when showerrors is true
If this whitelist is not used then all errors are shown if showerrors
is true. You can use this new option to explicitly allow backtraces
and descriptions to be shown to the user for only select error events.

If you provide a list of errors to show then anything not on that list
will not be shown to the user. The error will be logged etc as normal.

This was raised in
https://github.com/simplesamlphp/simplesamlphp/pull/2513